ImageSearchEngine/pyproject.toml
[project]
name = "imagesearchengine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.2",
    "pillow>=12.1.1",
    "qdrant-client>=1.17.0",
    "requests>=2.32.5",
    "sentence-transformers>=4.1,<5",  # important, fix version
    "transformers>=4.45,<5",
    "timm>=1.0.25",
    "torch>=2.5.1",
    "flask>=3.1.3",
]

[tool.uv.sources]
torch = {index = 'torch_cuda121'}
torchvision = {index = 'torch_cuda121'}

[[tool.uv.index]]
name = 'torch_cuda121'
url = "https://download.pytorch.org/whl/cu121"
explicit = true

